added accel_switch.hpp
