+++ sph/include/sph/hydro_std/density.hpp
@@ -37,7 +37,7 @@
 #include "cstone/findneighbors.hpp"
 #include "density_kern.hpp"
 #include "sph/sph.cuh"
-#include "sph/traits.hpp"
+#include "sph/particles_data_stubs.hpp"
 
 namespace sph
 {
