+++ sph/include/sph/hydro_std/momentum_energy.hpp
@@ -36,7 +36,7 @@
 #include "sph/kernels.hpp"
 #include "momentum_energy_kern.hpp"
 #include "sph/sph.cuh"
-#include "sph/traits.hpp"
+#include "sph/particles_data_stubs.hpp"
 
 namespace sph
 {
