+++ sph/include/sph/hydro_ve/av_switches.hpp
@@ -33,7 +33,7 @@
 
 #include "av_switches_kern.hpp"
 #include "sph/sph.cuh"
-#include "sph/traits.hpp"
+#include "sph/particles_data_stubs.hpp"
 
 namespace sph
 {
