+++ sph/include/sph/hydro_ve/iad_divv_curlv.hpp
@@ -34,7 +34,7 @@
 #include "iad_kern.hpp"
 #include "divv_curlv_kern.hpp"
 #include "sph/sph.cuh"
-#include "sph/traits.hpp"
+#include "sph/particles_data_stubs.hpp"
 
 namespace sph
 {
