+++ sph/include/sph/hydro_ve/ve_def_gradh.hpp
@@ -33,7 +33,7 @@
 
 #include "ve_def_gradh_kern.hpp"
 #include "sph/sph.cuh"
-#include "sph/traits.hpp"
+#include "sph/particles_data_stubs.hpp"
 
 namespace sph
 {
